Login change
During the restructure for multiple upload, moved login information from db_connect to clean_and_merge. The consequence of this is that users should only have to login once, at the start of running the upload process. This does mean that login errors will be delayed a bit, as the passwords are not used immediately when requested.